Added optional line to param file for time step
New script looks for 2 extra lines in param file that show the time step (as a fraction of days). If there is no such line, it defaults to 1.

Exp.tUse creates a vector of sequential times based on the aforementioned timestep. This allows for using input files with timesteps smaller than 1 day.