Edits to allow for AirSWOT run
Changed the layout of params file to allow for non-sequential algorithm inversion window (time-step issue).

Updated ReadParams so that params file now includes indices of all time steps in the inversion window. Also ensured a Bjerklie option if not specified in params file.

Updated SelObs to account for change in Exp time step vector, and ensure that dt value is correct (based on DAll.t, not Exp.tUse, which is an index).

In ProcessPrior, changed A0 init (sketchy but it was zeroing out for some reason, I think because dA for one reach never went negative).